+++ script/build_dev_js.sh
@@ -5,4 +5,4 @@ source ./script/utils.sh
 version=$(get_version)
 printf "Current frontend version: %s" "$version"
 
-build_js_dev_and_copy "$version"
+build_dev_js_and_copy "$version"