+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 * Add instruction for running all services for frontend within tmux session
 * Update README
 * Add wrappers for docker command
-* Use 'direnv' for managing env vars.
+* Use '.env' for managing env vars.
+* Refactor file structure of scripts and docs.
 
 ## 0.1.1.5